Move functional tests to functional/ directory.
