+++ tests/functional/BaseTestCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Tests;
+namespace Tests\Functional;
 
 class BaseTestCase extends \PHPUnit_Framework_TestCase
 {
@@ -35,21 +35,21 @@ class BaseTestCase extends \PHPUnit_Framework_TestCase
         $response = new \Slim\Http\Response();
 
         // Use the application settings
-        $settings = require __DIR__ . '/../src/settings.php';
+        $settings = require __DIR__ . '/../../src/settings.php';
 
         // Instantiate the application
         $app = new \Slim\App($settings);
 
         // Set up dependencies
-        require __DIR__ . '/../src/dependencies.php';
+        require __DIR__ . '/../../src/dependencies.php';
 
         // Register middleware
         if ($this->withMiddleware) {
-            require __DIR__ . '/../src/middleware.php';
+            require __DIR__ . '/../../src/middleware.php';
         }
 
         // Register routes
-        require __DIR__ . '/../src/routes.php';
+        require __DIR__ . '/../../src/routes.php';
 
         // Process the application
         $response = $app->process($request, $response);
